Updates to Elasticsearch VectoreStore Functionality.
