+++ packages/components/credentials/ElectricsearchUserPassword.credential.ts
@@ -15,13 +15,18 @@ class ElasticSearchUserPassword implements INodeCredential {
             'Refer to <a target="_blank" href="https://www.elastic.co/guide/en/kibana/current/tutorial-secure-access-to-kibana.html">official guide</a> on how to get User Password from ElasticSearch'
         this.inputs = [
             {
+                label: 'Cloud ID',
+                name: 'cloudId',
+                type: 'string'
+            },
+            {
                 label: 'ElasticSearch User',
-                name: 'elasticSearchUser',
+                name: 'username',
                 type: 'string'
             },
             {
                 label: 'ElasticSearch Password',
-                name: 'elasticSearchPassword',
+                name: 'password',
                 type: 'password'
             }
         ]
